requirements.txt
numpy
pandas
Flask
scikit-learn
gunicorn
pyYAML
-e .
